api/requirements.txt
fastapi[standard]==0.112.0
uvicorn[standard]
python-dotenv
pydantic
psycopg2-binary
SQLAlchemy
